Clippy

+++ src/metrics_classification.rs
@@ -613,7 +613,7 @@ mod tests {
         assert_abs_diff_eq!(x.accuracy(), 5.0 / 6.0_f32);
         assert_abs_diff_eq!(
             x.mcc(),
-            (2. * 3. - 1. * 0.) / (2.0f32 * 3. * 3. * 4.).sqrt() as f32
+            (2. * 3. - 1. * 0.) / (2.0f32 * 3. * 3. * 4.).sqrt()
         );
 
         assert_split_eq(
